requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.31.0
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.2.1
aiohttp>=3.8.6
websockets>=11.0.3
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
asyncpg>=0.29.0
